Add EIP-712 test
